refactor: move excalidraw-app outside src
