+++ src/tests/export.test.tsx
@@ -1,5 +1,5 @@
 import { render, waitFor } from "./test-utils";
-import ExcalidrawApp from "../excalidraw-app";
+import ExcalidrawApp from "../../excalidraw-app";
 import { API } from "./helpers/api";
 import {
   encodePngMetadata,
