+++ src/tests/zindex.test.tsx
@@ -1,6 +1,6 @@
 import ReactDOM from "react-dom";
 import { render } from "./test-utils";
-import ExcalidrawApp from "../excalidraw-app";
+import ExcalidrawApp from "../../excalidraw-app";
 import { reseed } from "../random";
 import {
   actionSendBackward,
